src/nexadash-backend-master/requirements.txt
blinker==1.4
Blueprints==2.3.0.2
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-PyMemcache==0.0.5
Flask-WTF==0.13.1
gevent==1.1.2
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
jwt==0.3.2
MarkupSafe==0.23
pycrypto==2.6.1
PyJWT==1.4.2
pymemcache==1.4.0
pymongo==3.3.1
python-slugify==1.2.1
requests==2.12.3
six==1.10.0
SQLAlchemy==1.1.4
Unidecode==0.4.19
Werkzeug==0.11.11
WTForms==2.1
